Task3/consumers/requirements.txt
kafka-python
py7zr
nltk
scikit-learn==0.24.1
pandas
cassandra-driver
twelvedata
websocket